+++ app/src/main/java/com/jdp/drivescoretechnicaltest/yourcreditscore/domain/model/YourCreditScoreDomainModel.kt
@@ -1,6 +1,6 @@
 package com.jdp.drivescoretechnicaltest.yourcreditscore.domain.model
 
 data class YourCreditScoreDomainModel(
-    val creditScore: Double,
-    val maxCreditScore: Double
+    val creditScore: Int,
+    val maxCreditScore: Int
 )
